+++ litebird_sim/simulations.py
@@ -2100,6 +2100,7 @@ class Simulation:
         component: str = "tod",
         append_to_report: bool = False,
         rng_hierarchy: RNGHierarchy | None = None,
+        conv_K_to_SR: bool = False,
     ):
         """A method to apply non-linearity to the observation.
 
@@ -2107,6 +2108,9 @@ class Simulation:
         :func:`.apply_quadratic_nonlin_to_observations()` that
         applies non-linearity to a list of :class:`.Observation` instance. Random number generators are obtained from the detector-level layer. As default it uses
         the `dets_random` field of a :class:`.Simulation` object for this.
+
+        conv_K_to_SR (bool, optional) is a flag for temperature to spectral radiance
+        units conversion. Defaults to False.
         """
         if rng_hierarchy is None:
             rng_hierarchy = self.rng_hierarchy
@@ -2122,6 +2126,7 @@ class Simulation:
             user_seed=user_seed,
             component=component,
             dets_random=dets_random,
+            conv_K_to_SR=conv_K_to_SR,
         )
 
         if append_to_report and MPI_COMM_WORLD.rank == 0:
@@ -2138,6 +2143,7 @@ class Simulation:
             self.append_to_report(
                 markdown_template,
                 g=g,
+                conv_K_to_SR=conv_K_to_SR,
             )
 
     @_profile